+++ src/lib/glaciersettings_global.h
