+++ src/lib/glaciersettingsplugin.h
