+++ src/lib/imagesmodel.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 Chupligin Sergey <neochapay@gmail.com>
+ * Copyright (C) 2017-2022 Chupligin Sergey <neochapay@gmail.com>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Library General Public
@@ -16,33 +16,45 @@
  * the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
  * Boston, MA 02110-1301, USA.
  */
-
-#ifndef TIMEZONESMODEL_H
-#define TIMEZONESMODEL_H
+#ifndef IMAGESMODEL_H
+#define IMAGESMODEL_H
 
 #include <QObject>
 #include <QAbstractListModel>
 
-#include <timezoneinfo.h>
+#include "glaciersettings_global.h"
 
-class TimeZonesModel:public QAbstractListModel
+class GLACIERSETTINGS_EXPORT ImagesModel: public QAbstractListModel
 {
     Q_OBJECT
+
+    Q_PROPERTY(QString path READ path WRITE setPath NOTIFY pathChanged)
+    Q_PROPERTY(bool reqursive READ reqursive WRITE setReqursive NOTIFY reqursiveChanged)
+
 public:
-    explicit TimeZonesModel(QObject *parent = 0);
+    explicit ImagesModel(QObject *parent);
+    void init();
 
     int rowCount(const QModelIndex &parent = QModelIndex()) const;
     QVariant data(const QModelIndex &index, int role) const;
-    QHash<int, QByteArray> roleNames() const {return m_hash;}
+    QHash<int, QByteArray> roleNames() const {return hash;}
 
-    Q_INVOKABLE void search(QString zoneName);
+    QList<QString> imageList;
+
+    QString path(){return m_imagesDir;}
+    void setPath(QString path);
+
+    bool reqursive(){return m_reqursive;}
+    void setReqursive(bool reqursive);
+
+signals:
+    void pathChanged();
+    void reqursiveChanged();
 
 private:
-    QHash<int,QByteArray> m_hash;
-    QStringList m_roleNames;
-
-    TimeZoneInfo* m_tzInfo;
-    QList<TimeZoneInfo> m_zones;
+    QHash<int,QByteArray> hash;
+    QString m_imagesDir;
+    bool m_reqursive;
 };
 
-#endif // TIMEZONESMODEL_H
+#endif // IMAGESMODEL_H
