+++ src/lib/satellitemodel.h
@@ -1,3 +1,22 @@
+/*
+ * Copyright (C) 2017-2022 Chupligin Sergey <neochapay@gmail.com>
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Library General Public
+ * License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Library General Public License for more details.
+ *
+ * You should have received a copy of the GNU Library General Public License
+ * along with this library; see the file COPYING.LIB.  If not, write to
+ * the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301, USA.
+ */
+
 #ifndef SATELLITEMODEL_H
 #define SATELLITEMODEL_H
 
@@ -8,10 +27,9 @@
 #include <QtQml/QQmlParserStatus>
 #include <QtPositioning/QGeoSatelliteInfoSource>
 
-QT_FORWARD_DECLARE_CLASS(QTimer)
-QT_FORWARD_DECLARE_CLASS(QGeoSatelliteInfoSource)
+#include "glaciersettings_global.h"
 
-class SatelliteModel : public QAbstractListModel, public QQmlParserStatus
+class GLACIERSETTINGS_EXPORT SatelliteModel : public QAbstractListModel, public QQmlParserStatus
 {
     Q_OBJECT
     Q_PROPERTY(bool running READ running WRITE setRunning NOTIFY runningChanged)
@@ -87,7 +105,5 @@ private:
     bool singleRequestServed;
 
 };
-
-QML_DECLARE_TYPE(SatelliteModel)
 
 #endif // SATELLITEMODEL_H
