+++ src/lib/settingspluginhost.cpp
