+++ src/lib/settingspluginhost.h
