+++ src/lib/settingspluginmanager.cpp
