+++ src/lib/settingspluginmanager.h
